+++ src/App.jsx
@@ -1,7 +1,7 @@
+import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Dashboard from './pages/dashboard/dashboard';
-import React from 'react';
-import GardenAlbum from './pages/GardenAlbum'
+import GardenAlbum from './pages/GardenAlbum/GardenAlbum';
 import Actividades from './pages/Actividades';
 
 function App() {
@@ -11,7 +11,6 @@ function App() {
         <Route path="/" element={<Dashboard />} />
         <Route path="/activities" element={<Actividades />} />
         <Route path="/garden" element={<GardenAlbum />} />
-        {/* Puedes agregar más rutas aquí si luego tienes más páginas */}
       </Routes>
     </BrowserRouter>
   );
