--- src/Styles/Album.css
@@ -1,27 +1,0 @@
-
-
-.album-wrapper {
-    position: relative;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 60px 0;
-    z-index: 5; /* encima de fondo */
-  }
-  
-  .album-bg {
-    width: 800px; /* ajusta según el tamaño real del cuaderno */
-    height: auto;
-    
-    left: 50%;
-    transform: translate(-50%, -50%);
-    z-index: 0;
-  }
-  
-  .album-content {
-    position: relative;
-    display: flex;
-    width: 800px; /* mismo tamaño que la imagen del cuaderno */
-    z-index: 1; /* encima de la imagen */
-  }
-
