--- src/Styles/NavBar.css
@@ -1,49 +1,0 @@
-.navbar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 20px;
-    position: fixed; /* ← importantísimo */
-    top: 0;
-    left: 0;
-    right: 0;
-    height: 70px;
-    background-color: #B8F2FE;
-    box-shadow: 0px 4px 6px rgba(160, 156, 156, 0.25);
-    z-index: 1000;
-  }
-  
-
-  .navbar-buttons {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 20px;
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    height: 70px;
-    background-color: #B8F2FE;
-    box-shadow: 0px 4px 6px rgba(160, 156, 156, 0.25);
-    z-index: 1000; /* asegura que esté encima de todo */
-  }
-  
-
-  .nav-left,
-  .nav-right {
-    display: flex;
-    gap: 12px;
-  }
-  
-  .nav-icon {
-    height: 40px;
-    transition: transform 0.2s ease, filter 0.2s ease;
-    cursor: pointer;
-  }
-  
-  .nav-icon:hover {
-    transform: scale(1.1);
-    filter: brightness(1.2) drop-shadow(0 0 5px rgba(0, 0, 0, 0.2));
-  }
-
