+++ src/components/Album/Album.css
@@ -1,0 +1,45 @@
+.album-wrapper {
+    position: relative;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 60px 0;
+    z-index: 5; 
+  }
+  
+  .album-bg {
+    width: 800px; 
+    height: auto;
+    
+    left: 50%;
+    transform: translate(-50%, -50%);
+    z-index: 0;
+  }
+  
+  .album-content {
+    position: relative;
+    display: flex;
+    width: 800px; 
+    z-index: 1;.garden-album {
+        position: relative;
+        max-height: 100%;
+        overflow-y: hidden;
+        overflow-x: hidden;
+        display: flex;
+        flex-direction: column;
+        justify-content: left;
+        padding-top: 0;
+      }
+      
+      .album-bg {
+        position: absolute;
+        top: 40%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+        width: 800px;
+        z-index: 1;
+        pointer-events: none;
+      }
+      
+  }
+
