+++ src/components/PageLeft/PageLeft.css
@@ -1,0 +1,72 @@
+.btns-wrapper {
+    display: flex;
+    justify-content: center;
+    gap: 3rem;
+    margin-top: 2rem;
+  }
+  
+  .btn-containerplus {
+    position: relative;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    right: 15%;
+    top: 130px;
+  }
+  
+  .btn-label {
+    font-family: 'Fredoka One', sans-serif;
+    color: #0E6229;
+    font-size: 1.2rem;
+    margin-top: 0.5rem;
+  }
+  
+  .img-btn {
+    position: relative;
+    display: flex;
+    justify-content: center;
+    left: 0%;
+    width: 200px;
+    height: auto;
+    transition: transform 0.2s ease;
+    z-index: 1;
+  }
+  
+  .grid-wrapper {
+    position: relative;
+    top: 100px;
+    right: 190px;
+  }
+  
+  .grid-3-columns {
+    display: grid;
+    grid-template-columns: repeat(3, 90px);
+    gap: 15px;
+    justify-content: center;
+    align-items: center;
+    padding: 10px 0;
+  }
+  
+  .grid-item {
+    width: 80px;
+    height: 80px;
+    border: 2px dashed #0E6229;
+    border-radius: 10px;
+    background-color: rgba(255, 255, 255, 0.7);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    transition: transform 0.2s ease;
+  }
+  
+  .grid-item:hover {
+    transform: scale(1.05);
+    cursor: pointer;
+  }
+  
+  .grid-item.selected {
+    border: 3px solid #0E6229;
+    background-color: rgba(255, 255, 255, 1);
+  }
+
