+++ src/components/ecoButton/ecoButton.css
@@ -12,6 +12,7 @@
     0 4px 0 #EBB372, 
     0 5px 10px rgba(0, 0, 0, 0.2);
     transition: transform 0.1s, box-shadow 0.1s, background-color 0.2s;
+    cursor: pointer;
 }
 
 .eco-button:active {
