+++ src/pages/GardenAlbum/GardenAlbum.css
@@ -15,7 +15,6 @@
   pointer-events: none;
 }
 
-/* Nubes (parte superior del cielo) */
 .img1 {
   top: 10px;
   left: 5%;
@@ -23,7 +22,6 @@
   height: auto;
 }
 
-/* Unión entre cielo y pasto */
 .img2 {
   bottom: 35%;
   left: 0;
@@ -31,7 +29,6 @@
   height: 200px;
 }
 
-/* Parte sobre el pasto */
 .img3 {
   bottom: 200px;
   left: 0;
