+++ src/pages/dashboard/dashboard.css
@@ -31,6 +31,7 @@
     gap: 20px;
     justify-content: space-around; /* Centra los elementos horizontalmente */
     align-items: center;
+    cursor: pointer;
 }
 
 .leftup{
